+++ History.md
@@ -209,5 +209,6 @@
 ======================
  * + corrected 'rbytes' property of incomplete files to 'size'
  * + added new config param 'maxFileSize'
+ * + added new config param 'serialzedFieldThreshold' for limiting the parsing of url encoded fields 
  * + corrected the file checksum value, now default value is 'null' when 'sha1sum' config param  is false, or if the file is incomplete
  * + resolved some bugs for the creation of the list of incomplete files